services/server/requirements.txt
numpy==2.0.0
pandas==2.2.3
python-dateutil==2.9.0
openpyxl
redis
Flask 
flask-wtf
Flask-SQLAlchemy==3.0.3
psycopg2-binary
flask-bcrypt
flask-login
Flask-Cors==6.0.0
Flask-Migrate==4.0.7
flask_jwt_extended
flask-caching
flask-session
python-dotenv
email_validator
PyJWT
scipy==1.14.1
requests==2.32.3
pyarrow
rq==1.16.2
scikit-learn
geopy
statsmodels
cmake
loess
pyinstaller
gunicorn==23.0.0
polars
xlsx2csv
fastexcel
python-logstash
pytest
pytest-flask
coverage
pytest-cov
pytest-docker
flask-socketio
# eventlet==0.38.0
gevent==24.2.1
pyodbc
pyyaml
ua-parser
user-agents
businessdate
affine>=2.4.0
dependency-injector>=4.41.0
numba>=0.56.4
pint
pillow>=10.2.0
pydantic>=2.4.2
pyproj>=3.5.0
shapely>=2.0.1
s3fs>=2022.1.0
zarr>=2.10.3
geopandas>=0.14.4
h3>=3.7.7
lmdbm>=0.0.6
fiona
cartopy
dask[dataframe]
# nptyping==2.4.1
xlrd
contextily
deprecation
h5py
netCDF4
xarray
pathos
pycountry
tqdm
sparse
xlsxwriter
pandas_datareader==0.10.0
cartopy
rasterio
ee==0.2
peewee==3.16.3
dataclasses
tabulate
paramiko==3.4.1
elastic-apm[flask]
netCDF4
bottleneck==1.4.2
ldap3==2.9.1
pycryptodome==3.21.0
celery==5.4.0
setuptools==78.1.1
urllib3==2.2.3
idna==3.10
fastparquet==2024.11.0
tzlocal==5.3.1
open-source-risk-engine==1.8.15.0
QuantLib==1.42.1